Migrate schenectady (EL30A fixed-width text report XLSX)
